refactor(ffi): split context and configuration entrypoints
